requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv
temporalio
gitpython
pytest
aiofiles
httpx
